+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added ability to invert train motor polarity (https://github.com/Novakasa/brickrail/issues/112)
+
 ### Fixed
 
 - Disable Godot HiDPI setting, since we currently don't react to user OS-level DPI scaling (https://github.com/Novakasa/brickrail/discussions/107#discussioncomment-5689736).